Moved each component to an external file. Some interactions were refined. Editing still doesn't work
